+++ package/lib/src/beamer_delegate.dart
@@ -22,7 +22,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
       'No longer used by this package, please remove any references to it. '
       'This feature was deprecated after v1.0.0.',
     )
-        this.preferUpdate = true,
+    this.preferUpdate = true,
     this.removeDuplicateHistory = true,
     this.notFoundPage = BeamPage.notFound,
     this.notFoundRedirect,
@@ -42,7 +42,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
       transitionDelegate: transitionDelegate,
     );
 
-    configuration = RouteInformation(location: initialPath);
+    configuration = RouteInformation(uri: Uri.parse(initialPath));
 
     updateListenable?.addListener(_update);
   }
@@ -564,10 +564,11 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
     bool replaceRouteInformation = false,
   }) {
     update(
-      configuration: RouteInformation(location: uri, state: routeState),
+      configuration: RouteInformation(uri: Uri.parse(uri), state: routeState),
       beamParameters: _currentBeamParameters.copyWith(
-        popConfiguration:
-            popToNamed != null ? RouteInformation(location: popToNamed) : null,
+        popConfiguration: popToNamed != null
+            ? RouteInformation(uri: Uri.parse(popToNamed))
+            : null,
         transitionDelegate: transitionDelegate ?? this.transitionDelegate,
         beamBackOnPop: beamBackOnPop,
         popBeamLocationOnPop: popBeamLocationOnPop,
@@ -623,7 +624,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
   }) {
     while (beamingHistory.isNotEmpty) {
       final index = beamingHistory.last.history.lastIndexWhere(
-        (element) => element.routeInformation.location == uri,
+        (element) => element.routeInformation.uri == Uri.parse(uri),
       );
       if (index == -1) {
         _disposeBeamLocation(beamingHistory.last);
@@ -793,12 +794,14 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
   @override
   SynchronousFuture<void> setInitialRoutePath(RouteInformation configuration) {
     _initialConfigurationReady = false;
-    final uri = Uri.parse(configuration.location ?? '/');
+    final uri = configuration.uri;
     if (currentBeamLocation is! EmptyBeamLocation) {
       configuration = currentBeamLocation.state.routeInformation;
     } else if (uri.path == '/') {
       configuration = RouteInformation(
-        location: initialPath + (uri.query.isNotEmpty ? '?${uri.query}' : ''),
+        uri: Uri.parse(
+          initialPath + (uri.query.isNotEmpty ? '?${uri.query}' : ''),
+        ),
       );
     }
     _deepLink ??= configuration.location;
@@ -807,12 +810,8 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
 
   @override
   SynchronousFuture<void> setNewRoutePath(RouteInformation configuration) {
-    if (!_initialConfigurationReady &&
-        configuration.location == '/' &&
-        (_deepLink != null || initialPath != '/')) {
-      configuration = configuration.copyWith(
-        location: _deepLink ?? initialPath,
-      );
+    if (configuration.uri.toString() == '/' && initialPath != '/') {
+      configuration = configuration.copyWith(location: initialPath);
     }
     update(configuration: configuration);
     return SynchronousFuture(null);
@@ -830,7 +829,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
     if (_parent == null) {
       if (!routeInformation.isEqualTo(_lastReportedRouteInformation)) {
         SystemNavigator.routeInformationUpdated(
-          location: routeInformation.location ?? '/',
+          uri: routeInformation.uri,
           state: routeInformation.state,
         );
         _lastReportedRouteInformation = routeInformation.copyWith();
@@ -905,7 +904,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
       );
     }
 
-    if (clearBeamingHistoryOn.contains(configuration.location)) {
+    if (clearBeamingHistoryOn.contains(configuration.uri.toString())) {
       _clearBeamingHistory();
     }
   }
@@ -966,7 +965,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
       _beamLocationCandidate = notFoundRedirect!;
     } else if (notFoundRedirectNamed != null) {
       _beamLocationCandidate = locationBuilder(
-        RouteInformation(location: notFoundRedirectNamed),
+        RouteInformation(uri: Uri.parse(notFoundRedirectNamed!)),
         _currentBeamParameters.copyWith(),
       );
     }
@@ -992,7 +991,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
       SystemChrome.setApplicationSwitcherDescription(
           ApplicationSwitcherDescription(
         label: _currentPages.last.title ??
-            currentBeamLocation.state.routeInformation.location,
+            currentBeamLocation.state.routeInformation.uri.path,
         primaryColor: Theme.of(context).primaryColor.value,
       ));
     }
@@ -1059,7 +1058,7 @@ class BeamerDelegate extends RouterDelegate<RouteInformation>
     if (_beamLocationCandidate is EmptyBeamLocation ||
         _beamLocationCandidate is NotFound) {
       update(
-        configuration: RouteInformation(location: initialPath),
+        configuration: RouteInformation(uri: Uri.parse(initialPath)),
         rebuild: false,
         updateParent: false,
         updateRouteInformation: false,